fix windows termlog & update health check test
